backend/requirements.txt
django
djangorestframework
google-generativeai
python-dotenv
django-cors-headers
